Refactor plot workflow and hardware dashboard
