+++ apps/api/pyproject.toml
@@ -18,6 +18,7 @@ dependencies = [
 dev = [
   "httpx>=0.27,<1.0",
   "pytest>=8.0,<9.0",
+  "ruff>=0.6,<1.0",
 ]
 
 [tool.setuptools]
@@ -28,3 +29,9 @@ where = ["src"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
+
+[tool.ruff]
+target-version = "py39"
+
+[tool.ruff.lint]
+select = ["F"]
