+++ apps/web/.eslintrc.cjs
@@ -1,0 +1,40 @@
+module.exports = {
+  root: true,
+  env: {
+    browser: true,
+    es2022: true,
+    node: true,
+  },
+  parser: "@typescript-eslint/parser",
+  parserOptions: {
+    ecmaVersion: "latest",
+    sourceType: "module",
+    ecmaFeatures: {
+      jsx: true,
+    },
+  },
+  plugins: ["@typescript-eslint", "react-hooks"],
+  ignorePatterns: [
+    "dist/",
+    "node_modules/",
+    "*.d.ts",
+    "*.js",
+    "*.tsbuildinfo",
+  ],
+  overrides: [
+    {
+      files: ["src/**/*.{ts,tsx}", "tests/**/*.{ts,tsx}", "vite.config.ts"],
+      rules: {
+        "no-unused-vars": "off",
+        "@typescript-eslint/no-unused-vars": [
+          "error",
+          {
+            argsIgnorePattern: "^_",
+            varsIgnorePattern: "^_",
+          },
+        ],
+        "react-hooks/rules-of-hooks": "error",
+      },
+    },
+  ],
+};
