+++ apps/web/src/features/plot-workflow/PlotWorkflowPanel.tsx
@@ -3,71 +3,19 @@ import type { ChangeEvent } from "react";
 import type { HardwareStatus, PlotterWorkspace } from "../../types/hardware";
 import type { PlotRun, PlotStageState } from "../../types/plotting";
 
+import {
+  formatPhysicalSize,
+  formatStageState,
+  getNestedRecord,
+  getRunStatusTone,
+  getStageLabel,
+  isRecord,
+} from "./plotWorkflowFormatters";
 import { usePlotWorkflow } from "./usePlotWorkflow";
 
 interface PlotWorkflowPanelProps {
   hardwareStatus: HardwareStatus;
   plotterWorkspace: PlotterWorkspace | null;
-}
-
-function getStageLabel(run: PlotRun | null) {
-  if (!run) {
-    return "No active run";
-  }
-  if (run.status === "plotting") {
-    return "Plotting";
-  }
-  if (run.status === "capturing") {
-    return "Capturing";
-  }
-  if (run.status === "completed") {
-    return "Completed";
-  }
-  if (run.status === "failed") {
-    return "Failed";
-  }
-  return "Preparing";
-}
-
-function getRunStatusTone(status: PlotRun["status"] | "idle") {
-  if (status === "completed") {
-    return "ok";
-  }
-  if (status === "pending" || status === "plotting" || status === "capturing") {
-    return "warn";
-  }
-  if (status === "failed") {
-    return "warn";
-  }
-  return "ok";
-}
-
-function formatStageState(stageState: PlotStageState) {
-  if (stageState.status === "in_progress") {
-    return "in progress";
-  }
-  return stageState.status;
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === "object" && value !== null;
-}
-
-function formatPhysicalSize(width: unknown, height: unknown) {
-  if (typeof width !== "number" || typeof height !== "number") {
-    return null;
-  }
-  return `${width.toFixed(1).replace(/\.0$/, "")} × ${height
-    .toFixed(1)
-    .replace(/\.0$/, "")} mm`;
-}
-
-function getNestedRecord(
-  record: Record<string, unknown> | null,
-  key: string,
-): Record<string, unknown> | null {
-  const value = record?.[key];
-  return isRecord(value) ? value : null;
 }
 
 export function PlotWorkflowPanel({
